Deleted and renamed some files, and import statements
