+++ package.json
@@ -94,7 +94,6 @@
     "dotenv": "8.2.0",
     "express": "4.17.1",
     "express-session": "1.17.0",
-    "lowdb": "1.0.0",
     "node-fetch": "2.6.0",
     "passport": "0.4.1",
     "passport-local": "1.0.0",
