--- server/database.js
@@ -1,7 +1,0 @@
-const pg = require('pg');
-
-const db = new pg.Pool({
-  connectionString: process.env.DATABASE_URL
-});
-
-module.exports = db;
