--- server/full-stack-project.example.conf
@@ -1,30 +1,0 @@
-server {
-
-  server_name full-stack-project.learningfuze.com;
-
-  root /home/ubuntu/$server_name/server/public;
-
-  # The following location block should be added if using React Router
-  #
-  # location / {
-  #   try_files $uri $uri/ /index.html;
-  # }
-
-  location /api {
-
-    # The following Host and X-Real-IP headers
-    # should be added if your Node.js back end needs them.
-    #
-    # proxy_set_header Host $host;
-    # proxy_set_header X-Real-IP $remote_addr;
-
-    # The following Upgrade and Connection headers
-    # should be added if proxying a websocket server.
-    #
-    # proxy_set_header Upgrade $http_upgrade;
-    # proxy_set_header Connection $connection_upgrade;
-
-    proxy_pass http://localhost:3001;
-  }
-
-}
