--- server/lowdb.js
@@ -1,6 +1,0 @@
-const low = require('lowdb');
-const FileSync = require('lowdb/adapters/FileSync');
-const adapter = new FileSync('db.json');
-const db = low(adapter);
-
-module.exports = db;
